Rename c/def files
