--- library/process/src/nativeInterop/cinterop/sys_pipe2.def
@@ -1,3 +1,0 @@
-package = io.matthewnelson.kmp.process.internal
-headers = sys_pipe2.h
-headerFilter = sys_pipe2.h
